pyproject.toml
[tool.poetry]
name = "openb3"
version = "0.1.0"
description = ""
authors = ["diogo2m <diogomonteiro.aluno@unipampa.edu.br>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^24.10.0"
yfinance = "^0.2.50"
matplotlib = "^3.9.2"
mysql-connector-python = "^9.1.0"
bcrypt = "^4.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
